==> real_estate/v6/01_clean_unsupervised_df.ipynb <==
import numpy as np
import pandas as pd
import pickle

from helpers import *

# Load data
df_ = pd.read_csv("~/real_estate/data/v4/unsupervised/yelp_unsupervised_df.csv")
df = df_.copy(deep=True)

# Load predictors, scaler
with open("../../real_estate/data/v6/predicting_nulls/scaler.pickle", "rb") as f:
    scaler = pickle.load(f)
with open("../../real_estate/data/v6/predicting_nulls/svr_pct_increasing.pickle", "rb") as f:
    svr_pct_increasing = pickle.load(f)  
with open("../../real_estate/data/v6/predicting_nulls/svr_pct_decreasing.pickle", "rb") as f:
    svr_pct_decreasing = pickle.load(f)  
with open("../../real_estate/data/v6/predicting_nulls/svr_median_sqft_value.pickle", "rb") as f:
    svr_median_sqft_value = pickle.load(f)
with open("../../real_estate/data/v6/predicting_nulls/svr_zhvi_singlefam.pickle", "rb") as f:
    svr_zhvi_singlefam = pickle.load(f)   
with open("../../real_estate/data/v6/predicting_nulls/svr_zri_sqft_value.pickle", "rb") as f:
    svr_zri_sqft_value = pickle.load(f)

# Scale columns
log_columns = ["zhvi_condo","people_count_lag","adult_count_lag","mean_adjusted_income_lag",
               "mean_adjusted_income_with_dep_lag","mean_real_estate_taxes_hh_lag","mean_mortgage_interest_hh_lag",
               "zhvi_singlefam","median_sqft_value","zri_sqft_value"]

for i in log_columns:
    df.loc[:,i] = log_with_zeros(df.loc[:,i].values)
    
scaling_columns = ['median_sqft_value','percent_decreasing','percent_increasing','zri_sqft_value','zhvi_condo',
                   'zhvi_singlefam','month','year','mean_rating','bar_review_count','restaurant_review_count',
                   'coffee_review_count','price_3_review_count','price_2_review_count','review_count',
                   'dependent_count_lag','sum_adjusted_income_lag','sum_real_estate_taxes_lag',
                   'sum_mortgage_interest_paid_lag','tax_return_count_lag','joint_return_count_lag','adult_count_lag',
                   'people_count_lag','mean_adjusted_income_lag','mean_real_estate_taxes_hh_lag',
                   'mean_mortgage_interest_hh_lag','mean_adjusted_income_with_dep_lag','90_day_treasury_bill_rate',
                   'mean_hourly_earnings_adj', 'mean_fixed_mortgage_rate','total_nonfarm_payroll']

# Scale the original df
df_scaled = scaler.transform(df[scaling_columns].values)
df_scaled = pd.DataFrame(df_scaled, columns=scaling_columns)

# Store indices of rows with null values across the four categories we're predicting
null_indices = {
    "pct_increasing": df_scaled.loc[df_scaled["percent_increasing"].isnull()].index,
    "pct_decreasing": df_scaled.loc[df_scaled["percent_decreasing"].isnull()].index,
    "median_sqft_value": df_scaled.loc[df_scaled["median_sqft_value"].isnull()].index,
    "zhvi_singlefam": df_scaled.loc[df_scaled["zhvi_singlefam"].isnull()].index,
    "zri_sqft_value": df_scaled.loc[df_scaled["zri_sqft_value"].isnull()].index,
}

columns_pred = [
    "90_day_treasury_bill_rate","adult_count_lag","bar_review_count","coffee_review_count", 
    "mean_adjusted_income_lag","mean_adjusted_income_with_dep_lag","mean_fixed_mortgage_rate",
    "mean_hourly_earnings_adj","mean_mortgage_interest_hh_lag","mean_rating","mean_real_estate_taxes_hh_lag",
    "people_count_lag","price_2_review_count","price_3_review_count","restaurant_review_count","review_count", 
    "total_nonfarm_payroll","zhvi_condo"]

pct_increasing_pred = svr_pct_increasing.predict(df_scaled.loc[null_indices["pct_increasing"],columns_pred].values)
pct_decreasing_pred = svr_pct_decreasing.predict(df_scaled.loc[null_indices["pct_decreasing"],columns_pred].values)
median_sqft_value_pred = svr_median_sqft_value.predict(df_scaled.loc[null_indices["median_sqft_value"],columns_pred].values)
zhvi_singlefam_pred = svr_zhvi_singlefam.predict(df_scaled.loc[null_indices["zhvi_singlefam"],columns_pred].values)
zri_sqft_value_pred = svr_zri_sqft_value.predict(df_scaled.loc[null_indices["zri_sqft_value"],columns_pred].values)

df_scaled_ = df_scaled.copy(deep=True)
df_scaled_.loc[null_indices["pct_increasing"],"percent_increasing"] = pct_increasing_pred
df_scaled_.loc[null_indices["pct_decreasing"],"percent_decreasing"] = pct_decreasing_pred
df_scaled_.loc[null_indices["median_sqft_value"],"median_sqft_value"] = median_sqft_value_pred
df_scaled_.loc[null_indices["zhvi_singlefam"],"zhvi_singlefam"] = zhvi_singlefam_pred
df_scaled_.loc[null_indices["zri_sqft_value"],"zri_sqft_value"] = zri_sqft_value_pred

# Rescale with minmax
df_rescaled = scaler.inverse_transform(df_scaled_)
df_rescaled = pd.DataFrame(df_rescaled, columns=scaling_columns)

# Rescale with log
log_columns_ = ["zhvi_condo","people_count_lag","adult_count_lag","mean_adjusted_income_lag",
               "mean_adjusted_income_with_dep_lag","mean_real_estate_taxes_hh_lag","mean_mortgage_interest_hh_lag",
               "zhvi_singlefam","median_sqft_value","zri_sqft_value"]

for i in log_columns_:
    df_rescaled[i] = np.exp(df_rescaled[i].values)

df_rescaled["postal_code"] = df_["postal_code"].values

# Export
df_rescaled.to_csv("~/real_estate/data/v6/semisupervised_df.csv",index=False)

df_rescaled.head()
