==> real_estate/v6/02_get_supervised_df.ipynb <==
import pandas as pd

### Yelp Data

df = pd.read_csv("~/real_estate/data/v6/semisupervised_df.csv")

features = [
    "median_sqft_value","percent_decreasing","percent_increasing","zri_sqft_value","zhvi_condo",
    "zhvi_singlefam","mean_rating","bar_review_count","restaurant_review_count","coffee_review_count",
    "price_2_review_count","price_3_review_count","review_count","dependent_count_lag",
    "sum_adjusted_income_lag","sum_real_estate_taxes_lag","sum_mortgage_interest_paid_lag",
    "tax_return_count_lag","joint_return_count_lag","adult_count_lag","people_count_lag",
    "mean_adjusted_income_lag","mean_real_estate_taxes_hh_lag","mean_mortgage_interest_hh_lag",
    "mean_adjusted_income_with_dep_lag","90_day_treasury_bill_rate","mean_hourly_earnings_adj",
    "mean_fixed_mortgage_rate","total_nonfarm_payroll"
]

target = "zhvi_condo"

# Store data from previous, future months and years
feature_mo = {}
feature_yr = {}
zhvi_condo_12_months = []

for i in features:
    feature_mo[i+"_mo"] = []
    feature_yr[i+"_yr"] = []

for i in df.index:
    
    current_row = df.iloc[i]
    region_name = current_row["postal_code"]
    month = int(current_row["month"])
    year = int(current_row["year"])
    next_year = year + 1
    
    # Get column values of previous time preiods
    last_year = year - 1
    if month == 1:
        last_month_month = 12
        last_month_year = year - 1
    else:
        last_month_month = month - 1
        last_month_year = year
    
    last_month_row = df.loc[(df["postal_code"] == region_name) & 
                            (df["month"] == last_month_month) &
                            (df["year"] == last_month_year)]
    
    last_year_row = df.loc[(df["postal_code"] == region_name) &
                           (df["month"] == month) &
                           (df["year"] == last_year)]
    
    next_year_row = df.loc[(df["postal_code"] == region_name) &
                           (df["month"] == month) &
                           (df["year"] == next_year)]
    
    for i in features:
        
        feature_mo[i+"_mo"].append(last_month_row[i].values)
        feature_yr[i+"_yr"].append(last_year_row[i].values)
        
    next_year_target = next_year_row["zhvi_condo"]
    
    if len(next_year_target) == 0:
        next_year_target = None
    else:
        next_year_target = next_year_target.values[0]
    
    zhvi_condo_12_months.append(next_year_target)

# Add data to dataframe
for i in feature_mo.keys():
    new_list = [j[0] if len(j) > 0 else None for j in feature_mo[i]]
    feature_mo[i] = new_list

for i in feature_yr.keys():
    new_list = [j[0] if len(j) > 0 else None for j in feature_yr[i]]
    feature_yr[i] = new_list
    
for i in feature_mo.keys():
    df[i] = feature_mo[i]
    
for i in feature_yr.keys():
    df[i] = feature_yr[i]
    
df["zhvi_condo_12_months"] = zhvi_condo_12_months

# Filter
df = df.dropna()

# Export
df.to_csv("~/real_estate/data/v6/yelp_supervised_df.csv", index=False)

print(df.shape)
df.head()
